+++ client/src/components/Table/styles.tsx
@@ -1,0 +1,22 @@
+import { styled } from "@mui/material/styles";
+import TableCell, { tableCellClasses } from "@mui/material/TableCell";
+import TableRow from "@mui/material/TableRow";
+
+export const StyledTableCell = styled(TableCell)(({ theme }) => ({
+	[`&.${tableCellClasses.head}`]: {
+		backgroundColor: theme.palette.common.black,
+		color: theme.palette.common.white
+	},
+	[`&.${tableCellClasses.body}`]: {
+		fontSize: 14
+	}
+}));
+
+export const StyledTableRow = styled(TableRow)(({ theme }) => ({
+	"&:nth-of-type(odd)": {
+		backgroundColor: theme.palette.action.hover
+	},
+	"&:last-child td, &:last-child th": {
+		border: 0
+	}
+}));
